feat：接入NumericComponent组件，规范化数值改变流程

+++ Server/Model/NKGMOBA/Battle/SkillSystem/Numeric/NumericWatcherComponent.cs
@@ -55,7 +55,7 @@ namespace ETModel
 			}
 		}
 
-		public void Run(NumericType numericType, long id, int value)
+		public void Run(NumericType numericType, long id, float value)
 		{
 			List<INumericWatcher> list;
 			if (!this.allWatchers.TryGetValue(numericType, out list))
